Move load aedt file
