+++ src/ansys/aedt/core/modules/material_lib.py
@@ -38,8 +38,8 @@ from ansys.aedt.core.generic.file_utils import open_file
 from ansys.aedt.core.generic.file_utils import read_json
 from ansys.aedt.core.generic.file_utils import write_configuration_file
 from ansys.aedt.core.generic.general_methods import pyaedt_function_handler
-from ansys.aedt.core.generic.load_aedt_file import load_entire_aedt_file
 from ansys.aedt.core.generic.settings import settings
+from ansys.aedt.core.internal.load_aedt_file import load_entire_aedt_file
 from ansys.aedt.core.modules.material import MatProperties
 from ansys.aedt.core.modules.material import Material
 from ansys.aedt.core.modules.material import SurfaceMaterial
