+++ src/pyaedt/generic/LoadAEDTFile.py
@@ -1,1 +1,1 @@
-from ansys.aedt.core.generic.load_aedt_file import *
+from ansys.aedt.core.internal.load_aedt_file import *
